+++ FE/src/app/admin_module/admin-routing.module.ts
@@ -4,7 +4,7 @@ import { ProductListComponent } from './component/product-list/product-list.comp
 import { NewProductComponent } from './component/new-product/new-product.component';
 import { AdminComponent } from './admin.component';
 import { OrdersListComponent } from './component/orders-list/orders-list.component';
-import { OrderInfosComponent } from '../shared/component/order-infos/order-infos.component';
+import { OrderInfosComponent } from '../shared_module/component/order-infos/order-infos.component';
 import { CategoryFormComponent } from './component/category-form/category-form.component';
 
 const routes: Routes = [
