--- FE/src/app/shared/component/paginator/paginator.component.spec.ts
@@ -1,16 +1,0 @@
-import { ComponentFixture, TestBed, async } from '@angular/core/testing';
-
-import { PaginatorComponent } from './paginator.component';
-import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule } from '@angular/forms';
-import { SharedModule } from 'src/app/shared/shared.module';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RouterTestingModule } from '@angular/router/testing';
-import { ProductCardComponent } from 'src/app/store_module/component/product-card/product-card.component';
-import { StoreHeaderComponent } from 'src/app/store_module/component/store-navbar/store-navbar.component';
-import { StoreComponent } from 'src/app/store_module/store.component';
-import { ProductFormComponent } from 'src/app/store_module/component/product-details/product-details.component';
-import { CheckOutComponent } from 'src/app/store_module/component/check-out/check-out.component';
-import { ShippingFormComponent } from 'src/app/store_module/component/shipping-form/shipping-form.component';
-
-describe('PaginatorComponent', () => {});
