+++ FE/src/app/shared_module/component/footer/footer.component.ts
