+++ FE/src/app/shared_module/component/header/header.component.scss
