+++ FE/src/app/shared_module/component/order-infos/order-infos.component.scss
