+++ FE/src/app/shared_module/component/paginator/paginator.component.scss
