+++ FE/src/app/shared_module/component/rating-bar/rating-bar.component.scss
