+++ FE/src/app/store_module/component/shipping-form/shipping-form.component.html
@@ -19,14 +19,7 @@
           <mat-label>phone number</mat-label>
           <input matInput placeholder="phone number..." type="tel" formControlName="phone" required>
         </mat-form-field>
-        <mat-form-field appearance="outline" class="address">
-          <mat-label>Select an address</mat-label>
-          <mat-select>
-            <mat-option value="test">
-              test
-            </mat-option>
-          </mat-select>
-        </mat-form-field>
+        
         <mat-form-field appearance="outline" class="address">
           <mat-label>address first line</mat-label>
           <input matInput placeholder="address first line..." formControlName="firstLine" required>
