+++ FE/src/app/store_module/store.component.scss
@@ -8,6 +8,7 @@
   display: flex;
   flex-wrap: wrap;
   justify-content: space-evenly;
+  font-family: 'Architects Daughter', cursive;
 }
 
 product-card {
