+++ FE/src/app/store_module/store.module.ts
@@ -7,7 +7,7 @@ import { StoreComponent } from './store.component';
 import { ProductDetails } from './component/product-details/product-details.component';
 import { CheckOutComponent } from './component/check-out/check-out.component';
 import { ShippingFormComponent } from './component/shipping-form/shipping-form.component';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../shared_module/shared.module';
 
 import { HeaderFormaterPipe } from './pipes/header-formater/header-formater.pipe';
 import { MatDialogModule } from '@angular/material/dialog';
