+++ FE/src/app/user_module/user.module.ts
@@ -6,7 +6,7 @@ import { UserComponent } from './user.component';
 import { UserHeaderComponent } from './component/user-navbar/user-navbar.component';
 import { AccountComponent } from './component/account/account.component';
 import { MyOrdersComponent } from './component/my-orders/my-orders.component';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../shared_module/shared.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MyReviewsComponent } from './component/my-reviews/my-reviews.component';
